requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio==7.2.20
apache-airflow-providers-amazon
boto3
pandas
pyarrow
s3fs